refactor: move models to messaging directory
