+++ lib/features/messaging/extensions/group_authorities.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:webtrit_phone/l10n/l10n.dart';
-import 'package:webtrit_phone/models/chat_member.dart';
+import 'package:webtrit_phone/models/messaging/chat_conversation/chat_member.dart';
 
 extension GroupAuthoritiesL10n on GroupAuthorities {
   String nameL10n(BuildContext context) {
