+++ lib/models/messaging/chat_conversation/chat_message_sync_cursor.dart
@@ -17,5 +17,7 @@ class ChatMessageSyncCursor extends Equatable {
   List<Object?> get props => [chatId, cursorType, time];
 
   @override
-  bool get stringify => true;
+  String toString() {
+    return 'ChatMessageSyncCursor(chatId: $chatId, cursorType: $cursorType, time: $time)';
+  }
 }
