+++ lib/models/messaging/chat_conversation/chat_outbox_read_cursor_entry.dart
@@ -25,4 +25,9 @@ class ChatOutboxReadCursorEntry extends Equatable {
 
   @override
   List<Object> get props => [chatId, time, sendAttempts];
+
+  @override
+  String toString() {
+    return 'ChatOutboxReadCursorEntry(chatId: $chatId, time: $time, sendAttempts: $sendAttempts)';
+  }
 }
