+++ lib/models/messaging/sms_conversation/sms_outbox_read_cursor_entry.dart
@@ -25,4 +25,9 @@ class SmsOutboxReadCursorEntry extends Equatable {
 
   @override
   List<Object> get props => [conversationId, time, sendAttempts];
+
+  @override
+  String toString() {
+    return 'SmsOutboxReadCursorEntry(conversationId: $conversationId, time: $time, sendAttempts: $sendAttempts)';
+  }
 }
