+++ lib/models/models.dart
@@ -1,20 +1,9 @@
-export 'active_message_push.dart';
 export 'agreement_status.dart';
 export 'audio_processing_settings.dart';
 export 'auth/auth.dart';
 export 'call/call.dart';
 export 'call_log_entry.dart';
 export 'call_to_action.dart';
-export 'chat.dart';
-export 'chat_member.dart';
-export 'chat_message.dart';
-export 'chat_message_read_cursor.dart';
-export 'chat_message_sync_cursor.dart';
-export 'chat_outbox_message_delete_entry.dart';
-export 'chat_outbox_message_edit_entry.dart';
-export 'chat_outbox_message_entry.dart';
-export 'chat_outbox_read_cursor_entry.dart';
-export 'chats_events.dart';
 export 'contact.dart';
 export 'contact_email.dart';
 export 'contact_phone.dart';
@@ -35,6 +24,7 @@ export 'local_contact_email.dart';
 export 'local_contact_phone.dart';
 export 'login_flavor.dart';
 export 'main_flavor.dart';
+export 'messaging/messaging.dart';
 export 'negotiation_settings.dart';
 export 'peer_connection_settings.dart';
 export 'recent.dart';
@@ -43,17 +33,9 @@ export 'repository/repository.dart';
 export 'rtp_codec_profile.dart';
 export 'session_status.dart';
 export 'settings_flavor.dart';
-export 'sms_conversation.dart';
-export 'sms_events.dart';
-export 'sms_message.dart';
-export 'sms_message_read_cursor.dart';
-export 'sms_message_sync_cursor.dart';
-export 'sms_outbox_message_delete_entry.dart';
-export 'sms_outbox_message_entry.dart';
-export 'sms_outbox_read_cursor_entry.dart';
 export 'system_info/system_info.dart';
+export 'system_notification.dart';
+export 'system_notification_event.dart';
+export 'system_notification_outbox_entry.dart';
 export 'video_capturing_settings.dart';
 export 'voicemail/voicemail.dart';
-export 'system_notification.dart';
-export 'system_notification_outbox_entry.dart';
-export 'system_notification_event.dart';
